+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add receiver for lib calls on right of pipe
 - add receiver for top lib on right of pipe
 - add `set/range` option support for `ui.pick`
+- fix hint sort for lib functions
 
 ## [0.13.7]
 - introduce `%{}` for explicit scopes